// src/enum/generic-error.enum.ts
export enum GenericErrorEnum {
    unknownError = "FoodieOrderProcessApi-Generic-0",
    invalidParameter = "FoodieOrderProcessApi-Generic-1",
    requestTimeout = "FoodieOrderProcessApi-Generic-2",
    apiFailure = "FoodieOrderProcessApi-Generic-3",
    forbiddenResource = "FoodieOrderProcessApi-Generic-4",
    payloadTooLarge = "FoodieOrderProcessApi-Generic-5",
    notFound = "FoodieOrderProcessApi-Generic-6",
    conflict = "FoodieOrderProcessApi-Generic-7",
}